Add possibility to watch JSON
It is now possible to watch the JSON code of tutorial examples.
Like this it is easier to understand how the tutorial has been made and
to reproduce them.

It displays the full tutorial and the current step.

The displayed window is movable and resizable in order to avoid to be
over important item on the screen.